move toValue from config/formats/utils.ts -> jsonToNestedValue in config/utils/jsonToNestedValue.ts
