+++ config/formats/javascript-commonJs.ts
@@ -1,7 +1,7 @@
 
 import StyleDictionary from 'style-dictionary';
 import { format } from "prettier";
-import { toValue } from './utils';
+import { jsonToNestedValue } from '../utilities/jsonToNestedValue';
 
 const { fileHeader } = StyleDictionary.formatHelpers;
 
@@ -10,7 +10,7 @@ export const javascriptCommonJs: StyleDictionary.Formatter = ({ dictionary, file
   let tokens = prefix ? { [prefix]: dictionary.tokens } : dictionary.tokens
   //
   const output = fileHeader({ file }) +
-    `module.exports = ${JSON.stringify(toValue(tokens), null, 2)}\n`
+    `module.exports = ${JSON.stringify(jsonToNestedValue(tokens), null, 2)}\n`
   // return prettified
   return format(output, { parser: "typescript", printWidth: 500 })
 }
